requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
biopython==1.86
bitsandbytes==0.49.1
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
comm==0.2.3
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
datasets==4.5.0
debugpy==1.8.20
decorator==5.2.1
dill==0.4.0
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonpickle==4.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
logistro==2.0.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.16.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
orjson==3.11.7
packaging==26.0
pandas==2.3.3
parso==0.8.6
peft==0.18.1
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
plotly==6.5.2
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydicom==3.0.1
Pygments==2.19.2
pyparsing==3.3.2
pyreadr==0.5.4
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
stack-data==0.6.3
streamlit==1.54.0
sympy==1.14.0
tcia_utils==3.2.1
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.1.0
triton==3.6.0
trl==0.28.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.22.0
